requirements-dev.txt
-r requirements.txt
pytest==9.0.3
pytest-flask==1.3.0
mypy==1.9.0
ruff==0.3.4
